chore(elastic-upgrade): reuse search types between services
