+++ packages/commons/src/index.ts
@@ -13,3 +13,4 @@ export * from './uuid'
 export * from './documents'
 export * from './http'
 export * from './logger'
+export * from './search'
