+++ packages/search/src/elasticsearch/utils.ts
@@ -23,76 +23,15 @@ import {
   SavedBundle,
   SavedOffice,
   SavedPractitioner,
-  SavedTask
+  SavedTask,
+  SearchDocument,
+  validStatusMapping,
+  IOperationHistory
 } from '@opencrvs/commons/types'
 import { findName } from '@search/features/fhir/fhir-utils'
 
-export const enum EVENT {
-  BIRTH = 'Birth',
-  DEATH = 'Death',
-  MARRIAGE = 'Marriage'
-}
-
-export const IN_PROGRESS_STATUS = 'IN_PROGRESS'
-export const ARCHIVED_STATUS = 'ARCHIVED'
-export const DECLARED_STATUS = 'DECLARED'
-export const REJECTED_STATUS = 'REJECTED'
-export const VALIDATED_STATUS = 'VALIDATED'
-const WAITING_VALIDATION_STATUS = 'WAITING_VALIDATION'
-export const REGISTERED_STATUS = 'REGISTERED'
-const REINSTATED_STATUS = 'REINSTATED'
-export const CERTIFIED_STATUS = 'CERTIFIED'
-export const ISSUED_STATUS = 'ISSUED'
-const REQUESTED_CORRECTION_STATUS = 'REQUESTED_CORRECTION'
-const DECLARATION_UPDATED_STATUS = 'DECLARATION_UPDATED'
-
 export const NOTIFICATION_TYPES = ['birth-notification', 'death-notification']
 export const NAME_EN = 'en'
-
-const validStatusMapping = {
-  [ARCHIVED_STATUS]: [
-    DECLARED_STATUS,
-    REJECTED_STATUS,
-    VALIDATED_STATUS
-  ] as const,
-  [IN_PROGRESS_STATUS]: [null] as const,
-  [DECLARED_STATUS]: [ARCHIVED_STATUS, null] as const,
-  [REJECTED_STATUS]: [
-    DECLARED_STATUS,
-    IN_PROGRESS_STATUS,
-    WAITING_VALIDATION_STATUS,
-    VALIDATED_STATUS,
-    ARCHIVED_STATUS
-  ] as const,
-  [VALIDATED_STATUS]: [
-    DECLARED_STATUS,
-    IN_PROGRESS_STATUS,
-    REJECTED_STATUS,
-    ARCHIVED_STATUS,
-    DECLARATION_UPDATED_STATUS,
-    null
-  ] as const,
-  [WAITING_VALIDATION_STATUS]: [
-    null,
-    DECLARED_STATUS,
-    IN_PROGRESS_STATUS,
-    REJECTED_STATUS,
-    VALIDATED_STATUS,
-    DECLARATION_UPDATED_STATUS
-  ] as const,
-  [REGISTERED_STATUS]: [
-    null,
-    DECLARED_STATUS,
-    IN_PROGRESS_STATUS,
-    REJECTED_STATUS,
-    VALIDATED_STATUS,
-    WAITING_VALIDATION_STATUS
-  ] as const,
-  [CERTIFIED_STATUS]: [REGISTERED_STATUS, ISSUED_STATUS] as const,
-  [ISSUED_STATUS]: [CERTIFIED_STATUS] as const,
-  [REQUESTED_CORRECTION_STATUS]: [REGISTERED_STATUS, CERTIFIED_STATUS] as const,
-  [REINSTATED_STATUS]: [ARCHIVED_STATUS] as const
-}
 
 export interface ICorrection {
   section: string
@@ -106,47 +45,6 @@ export interface IAssignment {
   firstName: string
   lastName: string
   officeName: string
-}
-
-export interface IOperationHistory {
-  operationType: keyof typeof validStatusMapping
-  operatedOn: string
-}
-
-export interface SearchDocument {
-  compositionId: string
-  compositionType?: string
-  event?: EVENT
-  type?: string
-  informantType?: string
-  contactNumber?: string
-  contactEmail?: string
-  dateOfDeclaration?: string
-  trackingId?: string
-  registrationNumber?: string
-  eventLocationId?: string
-  eventJurisdictionIds?: string[]
-  eventCountry?: string
-  declarationLocationId?: string
-  declarationJurisdictionIds?: string[]
-  rejectReason?: string
-  rejectComment?: string
-  relatesTo?: string[]
-  childFirstNames?: string
-  childFamilyName?: string
-  childFirstNamesLocal?: string
-  motherFirstNames?: string
-  motherFamilyName?: string
-  motherDoB?: string
-  motherIdentifier?: string
-  childDoB?: string
-  childIdentifier?: string
-  createdBy?: string
-  updatedBy?: string
-  createdAt?: string
-  modifiedAt?: string
-  assignment?: IAssignment | null
-  operationHistories?: IOperationHistory[]
 }
 
 export interface BirthDocument extends SearchDocument {
@@ -293,9 +191,7 @@ export async function detectDeathDuplicates(
 export async function getCreatedBy(compositionId: string) {
   const results = await searchByCompositionId(compositionId, client)
 
-  const result = results?.body?.hits?.hits[0]
-    ?._source as unknown as SearchDocument
-  return result?.createdBy
+  return results?.body?.hits?.hits[0]?._source?.createdBy
 }
 
 export const composeOperationHistories = (bundle: SavedBundle) => {
